Split Autodesk Service into different classes
