+++ app/jobs/upload_autodesk_files_job.rb
@@ -55,7 +55,7 @@ class UploadAutodeskFilesJob
 
   def perform_for_each_file(build, repository_id, filepath)
     autodesk_file = AutodeskFile.create!(repository_id:, filepath:)
-    autodesk_service = Autodesk.new(build)
+    autodesk_service = AutodeskFileUpload.new(build)
     urn = autodesk_service.upload_file_for_viewer(filepath)
 
     return if urn.nil?
